Add option to set inter-line gap.
This patch in a way goes together with the one preceding it.
It adds a new command line option (-gp) to set inter-line gap
when menu is displayed in vertical mode; this gap is controlled
by a new configuration parameter intlinegap.

Widget height is adjusted to provide space for these gaps, all
elements in the widget's first line are shifted down by an offset
of intlinegap/2 pixels to account for the upper half of inter-line
gap and, if menu is in vertical mode, we skip intliegap pixels
between each subsequent lines.