Add options to center window
Two new options are added, -xc and -yc to center menu window
in horizontal and vertical direction, respectively. If no
window width is specified on the command line, default width
spans the entire screen thus rendering horizontal centering
a noop; in that case, horizontal centering uses x offset,
if specified as total width of the marings.