Add support for translucent colors
For aesthetic reasons, it is sometimes desirable to overlay menu over
an existing image on screen using completely transparent or partially
translucent background color. In order to make this possible, we add
yet another way to specify color using syntax #AARRGGBB that contains
a value for an alpha channel.

On many X servers installed on hardware that support translucent colors,
they are not supported by default visual, usually 24-bit TrueColor.
So, before doing anything else, we need to find the right visual, and
use that instead of a default while creating menu window. In doing
this, we explicitly set border pixel and new private color map, as
we are no longer in position to rely on CopyFromParent (since parent
bit depth might not be the same). This also entails further changes
throught drw.c wherever window bit depth and color map are used.

After that, we are free to use translucent colors. Function
drw_clr_create is changed to produce translucent colors from
strings of the form #AARRGGBB; if X Server does not support
translucent colors, it discards AA part and returns an opaque
color.