Add option to draw a separator line
This line is drawn between the input line and displayed menu items
when menu is in vertical mode. New configuration parameteres include
line width, color and size of the padding on each side of the line.

This also involves ajusting menu height and shifting menu items down
to make space for the separator.