Add option to draw border around menu
To make space for the border, all elements get shifted
down and to the right, shrinking them at the same time
to fit reduced space left for displaying the menu;
this reduced space also affects menu item pagination in
function calcoffsets.

Border is drawn as a solid rectangular frame around
the menu window; we try to do it only when absolutely
necessary --- the first time around and when window
changes size with dynheight on; otherwise, border
that is already drawn on the PixMap is reused and
we only blank the space inside it.