Make dmenu work correctly under Wayland on GNOME DE
dmenu puts up an O-R window and relies on keyboard grab to
receive keyboard events. Under Wayland this works as long as
there is another X11 window and that X11 window has focus.
But, in practice, dmenu is often the only X11 window present!
Mutter does have a gsettings option to give keyboard focus
to selected O-R windows when requested (xwayland-allow-grabs
and xwayland-grab-access-rules). In order to go along with
this, this commit alters dmenu to issue keyboard grab request
on it's own window rather then on the root window and then
to request focus.